refactor venue to hall terminology and update related classes and documentation
