+++ application/src/main/java/org/mandarin/booking/app/hall/HallValidator.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.app.venue;
+package org.mandarin.booking.app.hall;
 
 public interface HallValidator {
     void checkHallExist(Long hallId);
