+++ application/src/main/java/org/mandarin/booking/app/hall/package-info.java
@@ -1,0 +1,4 @@
+@NamedInterface("hall")
+package org.mandarin.booking.app.hall;
+
+import org.springframework.modulith.NamedInterface;
