--- application/src/main/java/org/mandarin/booking/app/venue/package-info.java
@@ -1,4 +1,0 @@
-@NamedInterface("venue")
-package org.mandarin.booking.app.venue;
-
-import org.springframework.modulith.NamedInterface;
