+++ application/src/test/java/org/mandarin/booking/utils/TestConfig.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.mandarin.booking.adapter.TokenUtils;
+import org.mandarin.booking.app.hall.HallCommandRepository;
 import org.mandarin.booking.app.member.MemberCommandRepository;
 import org.mandarin.booking.app.show.ShowCommandRepository;
-import org.mandarin.booking.app.venue.HallCommandRepository;
 import org.mandarin.booking.domain.member.SecurePasswordEncoder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
