+++ domain/src/main/java/org/mandarin/booking/domain/show/ShowRegisterRequest.java
@@ -9,7 +9,7 @@ import org.mandarin.booking.domain.show.Show.Rating;
 import org.mandarin.booking.domain.show.Show.Type;
 
 public record ShowRegisterRequest(
-        @NotNull(message = "venue id is required")
+        @NotNull(message = "hall id is required")
         Long hallId,
 
         @NotBlank(message = "title is required")
